+++ sregistry/utils/__init__.py
@@ -9,6 +9,7 @@ from .terminal import (
     get_installdir,
     get_thumbnail,
     which,
+    confirm_action,
     confirm_delete
 )
 from .cache import get_cache