+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/singularityhub/sregistry-cli/tree/master) (0.0.x)
+ - allow anonymously accessing S3 buckets (0.2.24)
+   - add sregistry.utils.confirm_action
  - bugfix for rename (0.2.23)
  - Fixing sregistry search to honor a URI (0.2.22)
  - adding ability to specify working_dir for Google Cloud Build (0.2.21)